requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
google-auth==2.36.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.149.0
python-dateutil==2.9.0
requests==2.32.3
gunicorn==23.0.0
psycopg[binary]==3.2.11
pandas==2.2.3
openpyxl==3.1.2
numpy<2
openai==1.44.0
